Remove sample blog and update article tag styling
The sample blog was removed from the posts as it was used just for testing and is not needed for production use.

In addition, the method of applying styles to article tags was updated to improve flexibility by replacing hardcoded tags to a more programmatic approach using an index on the `textsStyles` array.

Finally, the `.idea` folder was added to `.gitignore` for JetBrains IDEs, and a new blog titled "How to fix Vscode asking for Github login on macOS" was introduced.

The `twitter` object was also removed from the `seoConfig` in `next-seo.config.ts` file as it is not used in the configuration.

In `sitemap.xml`, entries were updated with new metadata.