Update sitemap.xml and theme toggling logic
Updated sitemap.xml URLs according to the new domain and adjusted the last modification time. Also, refined the theme toggling condition in Header.tsx to improve performance. The condition now checks if the theme is not "light" instead of if it is "dark", which eliminates the need to check for two variables. This makes the site more friendly for search engines and improves UX as it allows more efficient theme toggling.